developer/django_practices/api/requirements.txt
Django==3.0.9
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
python-dotenv==0.13.0
PyYAML==5.3.1
uritemplate==3.0.1

